chore: move facing dir to entity
