+++ Assets/Animations/Customer/default_idle_down.anim
@@ -6,7 +6,7 @@ AnimationClip:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: default_walk_right
+  m_Name: default_idle_down
   serializedVersion: 7
   m_Legacy: 0
   m_Compressed: 0
@@ -21,19 +21,19 @@ AnimationClip:
   - serializedVersion: 2
     curve:
     - time: 0
-      value: {fileID: -5004526041978996066, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
-    - time: 0.16666667
-      value: {fileID: 6580674900253841023, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
-    - time: 0.33333334
-      value: {fileID: 812993532842763001, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
+      value: {fileID: 9079905429464953686, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
+    - time: 0.25
+      value: {fileID: -5883383104323906680, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
     - time: 0.5
-      value: {fileID: 8802695626040928466, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
+      value: {fileID: -7851985453260657375, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
+    - time: 0.75
+      value: {fileID: 3170977611123660100, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
     attribute: m_Sprite
     path: 
     classID: 212
     script: {fileID: 0}
     flags: 2
-  m_SampleRate: 6
+  m_SampleRate: 4
   m_WrapMode: 0
   m_Bounds:
     m_Center: {x: 0, y: 0, z: 0}
@@ -50,16 +50,16 @@ AnimationClip:
       isIntCurve: 0
       isSerializeReferenceCurve: 0
     pptrCurveMapping:
-    - {fileID: -5004526041978996066, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
-    - {fileID: 6580674900253841023, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
-    - {fileID: 812993532842763001, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
-    - {fileID: 8802695626040928466, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
+    - {fileID: 9079905429464953686, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
+    - {fileID: -5883383104323906680, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
+    - {fileID: -7851985453260657375, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
+    - {fileID: 3170977611123660100, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
   m_AnimationClipSettings:
     serializedVersion: 2
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.6666667
+    m_StopTime: 1
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
