+++ Assets/Animations/Customer/default_idle_left.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 93e0b9103ab88f349a40bf89fb8faffd
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
