+++ Assets/Animations/Customer/default_idle_right.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f36adb7fec578084a84cb27affbd07d0
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
