+++ Assets/Animations/Customer/default_idle_up.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 753f736a9bda5ee4ab93f4c2c40f5d9c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
