+++ Assets/Animations/Customer/default_walk_up.anim
@@ -17,21 +17,49 @@ AnimationClip:
   m_PositionCurves: []
   m_ScaleCurves: []
   m_FloatCurves: []
-  m_PPtrCurves: []
-  m_SampleRate: 60
+  m_PPtrCurves:
+  - serializedVersion: 2
+    curve:
+    - time: 0
+      value: {fileID: 886287592483588472, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
+    - time: 0.16666667
+      value: {fileID: 435636438829087750, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
+    - time: 0.33333334
+      value: {fileID: 1853100027989576181, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
+    - time: 0.5
+      value: {fileID: 518130849492762848, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
+    attribute: m_Sprite
+    path: 
+    classID: 212
+    script: {fileID: 0}
+    flags: 2
+  m_SampleRate: 6
   m_WrapMode: 0
   m_Bounds:
     m_Center: {x: 0, y: 0, z: 0}
     m_Extent: {x: 0, y: 0, z: 0}
   m_ClipBindingConstant:
-    genericBindings: []
-    pptrCurveMapping: []
+    genericBindings:
+    - serializedVersion: 2
+      path: 0
+      attribute: 0
+      script: {fileID: 0}
+      typeID: 212
+      customType: 23
+      isPPtrCurve: 1
+      isIntCurve: 0
+      isSerializeReferenceCurve: 0
+    pptrCurveMapping:
+    - {fileID: 886287592483588472, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
+    - {fileID: 435636438829087750, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
+    - {fileID: 1853100027989576181, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
+    - {fileID: 518130849492762848, guid: 38603c4ee08bf41b6952e928c30aade8, type: 3}
   m_AnimationClipSettings:
     serializedVersion: 2
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 1
+    m_StopTime: 0.6666667
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
