+++ Assets/Prefab/Customer.prefab
@@ -18,7 +18,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!4 &1301589784978832981
 Transform:
   m_ObjectHideFlags: 0
