+++ Assets/Scripts/Entity.cs
@@ -1,4 +1,5 @@
 using UnityEngine;
+using static Player;
 
 public class Entity : MonoBehaviour
 {
@@ -9,7 +10,10 @@ public class Entity : MonoBehaviour
     public Animator anim;
     public SpriteRenderer sr;
 
-    public StateMachine stateMachine { get; private set; }   
+    public StateMachine stateMachine { get; private set; }
+
+    public enum FacingDirection { Up, Down, Left, Right }
+    public FacingDirection facingDirection { get; private set; } = FacingDirection.Down;
 
 
     protected virtual void Awake()
@@ -20,5 +24,18 @@ public class Entity : MonoBehaviour
     protected virtual void Update()
     {
         stateMachine.CallUpdateCurrentState();
-    } 
+    }
+
+    protected void UpdateFacingDirection(Vector2 input)
+    {
+        if (input == Vector2.zero) return;
+
+        if (Mathf.Abs(input.x) >= Mathf.Abs(input.y))
+            facingDirection = input.x > 0 ? FacingDirection.Right : FacingDirection.Left;
+        else
+            facingDirection = input.y > 0 ? FacingDirection.Up : FacingDirection.Down;
+
+        sr.flipX = (facingDirection == FacingDirection.Right);
+
+    }
 }
