+++ Assets/Scripts/StateMachine/PlayerState.cs
@@ -13,10 +13,5 @@ public class PlayerState : EntityState
     {
         anim.SetFloat("xSpeed", player.moveInput.x);
         anim.SetFloat("ySpeed", player.moveInput.y);
-
-        anim.SetFloat("idleX", player.facingDirection == FacingDirection.Right ? 1 :
-                        player.facingDirection == FacingDirection.Left ? -1 : 0);
-        anim.SetFloat("idleY", player.facingDirection == FacingDirection.Up ? 1 :
-                                player.facingDirection == FacingDirection.Down ? -1 : 0);
     }
 }
